Call refresh token handler in auth fetch
Once the authenticated fetch has been built, it may on 401 use the
refresh token to get a new access token. In this case, it should also
call the refresh token handler in the case of a refresh token rotation,
so that the hook provided by the user is called and gets executed with
the latest refresh token.