Enforce calling handler on login with token
When using the refresh token as part of the login process, to get an
initial access token without requiring a user interaction, the refresh
token may be rotated. The handler was properly injected to the refresh
handler, but it wasn't properly used; this fixes it and adds a test to
make sure the handler is used.