Add refresh token handling hooks (#1284)
In order to allow for some use cases where using the storage isn't the
most practical solution, this makes it possible to run some custom code
when getting an access token. The `onNewRefreshToken` hook allows
to register a callback that gets invoked when a new refresh token is sent
by the Identity Provider. The callback is provided to the sesion constructor.

* Call refresh token handler in auth fetch

Once the authenticated fetch has been built, it may on 401 use the
refresh token to get a new access token. In this case, it should also
call the refresh token handler in the case of a refresh token rotation,
so that the hook provided by the user is called and gets executed with
the latest refresh token.

* Inject handler when getting session from storage

Users may get their session from the storage, and not from the
constructor. In this case, since the handler isn't serialized with the
session in storage, they'd have to provied the handler to the
getSessionFromStorage function in the same way they have to provide it
to the session constructor.

* Enforce calling handler on login with token

When using the refresh token as part of the login process, to get an
initial access token without requiring a user interaction, the refresh
token may be rotated. The handler was properly injected to the refresh
handler, but it wasn't properly used; this fixes it and adds a test to
make sure the handler is used.