Use iframes to handle redirect
This introduces the usage of iframes to handle redirection to the
Identity Provider, instead of redirecting the main window. The iframe
uses "prompt=none" in order to go through silent authentication, and to
be directly returned without any user interaction by the Identity
Provider. The iframe then posts the return IRI (which contains the
authorization code) to the parent window, which is able to use it as if
it has been redirected normally.

The main assumption here is that iframes can use cookies set on the
browser.

Note: There are still debug logs in this codebase, they'll be removed in
an upcoming PR.