Share single iframe reference
Instead of adding the iframe to the DOM, and retrieveing it though it
too, this creates the iframe node once, and adds/removes it from the DOM
as appropriate. The iframe no longer needs to be retrieved through a DOM
query, which means it may not have an id, which prevents from ever
colliding with an existing iframe the calling code may have put in the
DOM.

This also replaces mock calls by actual usage of jsdom: instead of
capturing calls to window to pretend an iframe exists or not, this
simply actually puts an iframe in jsdom and retrieves from it too.