Silent authentication in an iframe (works in Chrome)
This introduces the usage of iframes to handle redirection to the
Identity Provider, instead of redirecting the main window. The iframe
uses "prompt=none" in order to go through silent authentication, and to
be directly returned without any user interaction by the Identity
Provider. The iframe then posts the return IRI (which contains the
authorization code) to the parent window, which is able to use it as if
it has been redirected normally.

The main assumption here is that iframes can use cookies set on the
browser.

This introduces a new file gathering all iframe-related functions. These
are then simply called in the appropriate contexts across the codebase.
The introduced functions aime at:
- Redirecting the browser to an IRI in an iframe
- Letting the parent window know to what the redirect IRI the Solid Identity
Provider has  redirected the iframe to.
- The main window cleans up the iframe after it handled the redirection.
- The session triggers the iframe-based flow based on a signal. In
future commit, this signal may be emitted when an access token expires.

Important note: this has only been tested for Chrome.