Added dependency from oidc to core package (#1339)
Initially, `oidc-client-ext` was meant to be unrelated to anything
solid-specific, which is why it wasn't dependant on
`solid-client-authn-core`. However, the `core` package mostly containing
type definitions and shared utils functions between the browser and node
modules, it isn't a big dependency on `oidc-client-ext`, but it reduces
dependency, so the benefits outwheight the costs.

Also, migrating to josev@3 will increase the portion of shared code
related to crypto and dpop, making the split between core and oidc even
more painful.