Update to jose@3
This migrates the browser code to jose version 3 (but importing
from @inrupt/jose-legacy-modules, which is the same code but
importable from the top level so that it remains compatible with
tools that do not support package exports, like Webpack 4).

The contents of this commit is primarily authored by Nicolas; I
merely dredged through
https://github.com/inrupt/solid-client-authn-js/pull/1356/files,
skipped the upgrade to Webpack 5 and the Prettier changes, and
replaced jose imports by imports from @inrupt/jose-legacy-modules.

Co-authored-by: Zwifi <nseydoux@inrupt.com>