Extend session info storage with DPoP key
In the case of NodeJS server code, where the sessions aren't kept in
memory but stored and restored from storage, it is necessary to store
the DPoP key along the refresh token in case the latter is bound to the
former. While this is safe to do with NodeJS storage (or at least, not
less safe than storing the refresh token), it isn't recommended to store
the DPoP key in the browser.