Feat/node/2165 client credentials (#1550)
This implements a first version of the client credentials grant,
validated via manual testing. As for the refresh flow, if `session.login` is called with the appropriate parameters, the client credential flow is triggered and directly returns a valid session, i.e. no redirection is needed.

In order to implement the client credential flow handler, it is necessary to be able to discriminate between a dynamically registered client, a Solid-OIDC client, and a statically registered client, since only the latter may go through the client credential flow.

* Build an token type agnostic authenticated fetch

Instead of having two separate functions to build either a DPoP or a
Bearer fetch, only one functions calls to the appropriate underlying
fetch factory.

In particular, this makes test code in other parts of the codebase (e.g.
ClientCredentialsHandler) more compact.

The self-service client credentials endpoint expects basic auth, and not
client secret post.

Co-authored-by: Vincent <Vinnl@users.noreply.github.com>