Switch from jose submodules to legacy (#1561)
Using submodules import for jose creates interoperability issues with
parts of the ecosystem that are not up-to-speed yet with this mechanism,
e.g. jest. It is manageable internally, but it actually also impacts our
dependents, which we want to avoid as much as possible.

To work around the issue, all dependencies to jose go through an
intermediary packages which re-exports jose submodules as legacy
submodules.

Co-authored-by: Vincent <Vinnl@users.noreply.github.com>