Fix onError callback issue
This makes it possible to pass a callback all the way down to the error
handler. In particular, it:
- Makes the onError an option of the Session::handleIncomingRedirect
method, rather than an attribute of the Session
- Makes sure that the onError handler isn't passed as the onNewToken
handler in ClientAuthentication

This also prevents the onError callback to be passed to the iframe
context, since we're removing the iframe code soon it wasn't necessary
anymore.