Ask for offline scope in browser (#1580)
Asking for the `offline_scope` requests a refresh token from the OIDC provider. The client profile, either the metadata provided during DCR or the information in the client WebID document, must reflect that the client may ask for the `offline_access` scope at the authorization endpoint. The `prompt=consent` parameter must also be specified.