Properly handle client credentials (#1594)
Going through the refresh token flow with a dynamically registered
client requires providing a client ID and secret, which are the
conditions to trigger the client credentials flow. However, the client
credential flow only works for __statically__ registered clients, which
leads to a bug.

This adds test to ensure that the appropriate OIDC handler is called
depending on the provided parameters. These tests aren't great, because
they pretty much only test implementation rather than behaviour, and
they rely on internal undocumented API, but they cover for issues that
have bitten us a couple of times now.