Call callbacks on events (#1576)
This changes how events are handled and reported back to the developer of a consuming app. The `Session` class has several event-related hooks, such as `onLogin` or the newly introduced `onError`, where it is possibke register callbacks in order to run custom code when the event is fired.

This introduces an internal change: instead of passing the callbacks in the code, so that they are injected down to the level where they need to be called, we now pass around an `EventEmitter` instance (which happens to be the `Session` itself. The event emitter is used to emit events when appropriate, and the callbacks are registered to the event emitter to run when the event they are listening for happens.

Co-authored-by: Zwifi <nseydoux@inrupt.com>
Co-authored-by: dgbarclay <dan@danbarclay.com>
Co-authored-by: Dan Barclay <60509571+dgbarclay@users.noreply.github.com>