Proactive token refresh (#1590)
Instead of waiting for a request to fail to renew the access token, the
authenticated fetch now sets a timeout to trigger the refresh flow
before the expiration of the access token.

FIXME: unsatisfactory tests

Unfortunately, Jest mocks timer don't look like they are behaving the
way we'd intend. To be able to run our tests, it has been necessary to
use real timers instead. It makes the tests somewhat slower (because we
have to wait for an actual length of time).