Error handling on silent authentication (#1607)
When using refresh tokens to keep the session alive, for some reason the
silent authentication may fail. In this case, we want to notify the app
that it no longer is authenticated, and that it should prompt the user
for re-authentication.

The Session object exposes an expiration date to the app. This
expiration date is now updated each time the token is refreshed. This
also ensures that the session expiration callback is called
appropriately when the access token cannot be refreshed.