Use subject_type public during DCR
This fixes https://github.com/inrupt/solid-client-authn-js/issues/1551,
and endures that during Dynamic Client Registration, the
`subject_type` parameter is properly set to `public`, and not to
`pairwise`. Indeed, in Solid, the subject identifier is the WebID, which
is global and unique: it should be shared among all clients. Having the
OP generate a different subject claim for each client for a given
subject does not make sense in the context of Solid.