Fix authn when using client identifiers
When using a Client Identifier as defined in Solid-OIDC, a client does not have a secret. Based on https://datatracker.ietf.org/doc/html/rfc6749#section-2.3.1, it looks like it is acceptable to only provide a client id, but no secret, in the authentication information stored in the body of the request to the token endpoint.

However, this is currently running into an issue with the underlying OIDC library. A conversation is happening at https://github.com/panva/node-openid-client/discussions/430. In the meantime, an arbitrary secret is given to the library, which works around the issue and is ignored by the server.