Fix authn when using client identifiers (#1793)
When using a Client Identifier as defined in Solid-OIDC, a client does not have a secret. Based on https://datatracker.ietf.org/doc/html/rfc6749#section-2.3.1, it looks like it is acceptable to only provide a client id, but no secret, in the authentication information stored in the body of the request to the token endpoint.