No longer store the ID token in local storage
The ID token becoming a part of the authentication flow and being exchangeable for an Access token, it is now a critical piece of information and should be handled as such. It means, among other things, that it should not be stored in local storage. A future PR will keep it in the fetch closure, like the access token.