Remove node-fetch dependency
Our packages already depend on cross-fetch, which in turns depends on node-fetch. The dependency on node-fetch was only introduced in the tests because cross-fetch was mocked and it created unintended consequences that were not fully understood initially. Now, mocking of cross-fetch is done properly, which makes depending on node-fetch unecessary. cross-fetch comes with its own type definitions too.