Improve mocked types for fetch and core
Thanks to @matthieubosquet for looking into this, which replaces a cast to any into a proper typing.
The imported module is explicitly typed so that both the requireActual and requireMock statements can reuse the proper type instead of any. Note that the type cast in the initial `jest.mock("cross-fetch", () => { ... });` brings no additional type safety, but it removes some `as any` casts. Using the proposed syntax when requirering the created mock, however, does make mocking types safer.

This includes:
- fetchJwks in browser/ClientAuthn
- fetch in both core, browser and node modules
- the core module in the browser and node ones