Convert e2e tests to node-tap
The Jest runner has some subtles differences with the Node environment. It is fine for our unit tests, but we want our end-to-end tests to run in an environment as close as possible to the actual environment in which the code will actually run.

To this end, this experiments with node-tap as a test framework. There is no runner, the tests just run as regular node scripts.