No longer store the ID token in local storage
The ID token becoming a part of the authentication flow and being exchangeable for an Access token, it is now a critical piece of information and should be handled as such. It means, among other things, that it should not be stored in local storage. A future PR will keep it in the fetch closure, like the access token.

The ID token should no longer be stored, so it is no longer accepted as a parameter to the session info storage function. SessionInfoManager also no longer has any reference to ID tokens.