Add Node e2e workflow
Instead of having multiple environment variables with different names (e.g. one IDP for dev-next and one for prod), this makes it so that the end-to-end test only expect a single set of environment variables (e.e., one IDP only), but will run in different environments providing the appropriate values for these variables in each context.