Prevent silent auth cycle on authentication error
When an authentication error happens, local storage is now cleared in order to prevent further silent login attempts. This should fix the issue where failed login cycle endlessly.

In order to implement this, the Session object listens on authentication events bubbled up by the handlers. In particular, on login, the current session ID is stored in local storage to enable silent auth in the future, and it is cleared on logout/auth error.