Clear verifier after token exchange
The token verifier should only be used once when exchanging the authorization code at the token endpoint of the OIDC provider. Beyond this point, the code verifier should be cleared from memory.

In the browser case, this piece of data is stored in two places: the oidc-client specific storage in localStorage, and the solid-client-authn storage. The former is automatically cleaned up when using oidc-client for the token exchange, which only happens when requesting a "plain" bearer token, and not a dpop-bound one. In the dpop case, our custom code is performing the token exchange, which is why the oidc-client state needs to be manually cleared.