Prevent silent auth cycle on authentication error (#2002)
When an authentication error happens, local storage is now cleared in order to prevent further silent login attempts. This should fix the issue where failed login cycle endlessly.

In order to implement this, the Session object listens on authentication events bubbled up by the handlers. In particular, on login, the current session ID is stored in local storage to enable silent auth in the future, and it is cleared on logout/auth error.

To cleanup code organization, as much localStorage references as possible are moved at the top-level in the Session class.

Session now has an internal logout method which can either send a logout signal or not. The signal is only meant to be sent on user-initiated logout, but the function is also called in the case of auth error or session expiration.