Disable using the ESS session endpoint
The ESS session endpoint was a legacy workaround introduced in an experiment for long-running sessions using resource server-side cookies. This approach had multiple issues, and the experiment showed that other solutions were preferable.

This endpoint is disabled by default on ESS, and this workaround was opt-in, so disabling it should be transparent to all users, as to the best of our knowledge noone was still relying on this.