Support missing ID token in client credential flow
By definition, the client credential flow happens without the resource owner: the client acts on its own behalf. In some Solid-OIDC providers implementations, this is slightly worked around by binding the client to a resource owner account so that it still gets a WebID. However, these workaround are somewhat implementation-specific, and not all result in striclty similar behaviours. In particular, the ESS Solid-OIDC broker produces an ID token in response to a client credential grant token exchange, but that isn't the case of the CSS Solid-OIDC provider.

This commit clears our assumption that an ID token must always be present: instead, it is acknowledged that in the absence of an ID token, the access token should contain a `webid` claim, which is sufficient for our purpose. Note that this is a temporary fix: the authentication panel is moving away from an access token being present at all in the Solid-OIDC provider response. Dedicated work will be required to address the use case currently covered by the client credential flow. When the Solid-OIDC specification provides normative behavior there, this library should be updated accordingly.