Remove reference to window.localstorage from Session constructor
A temporary fix was put into place in
3fd7bb44de656bf044e9b09bb3f29fcada5e8de4 which cleared a localstorage
value set by a temporary ESS session restoration workflow. This caused
window to be accessed in the session constructor, which caused issues
with server-side rendering.

The localstorage value no longer needs to be cleared as it is unused,
so this line has been removed.