Try restoring client Identifier from storage
When logging a session back in after restoring it from storage, authentication fails if using a client identifier, while it works well if using dynamic client registration. The call to login from `getSessionFromStorage` does not include the client identifier, and even if it eventually should be retrieved properly from storage, at that point the client has been tagged dynamic, which probably causes the mismatch.

This is an initial commit to see if this resolves the issue, tests will come in a second time.