Remove InruptError class
The InruptError class is making use of the inrupt vocab library, which relies on some local storage behavior. This behavior breaks under jest 29, and since thiese features were not used in this library and should be implemented in a proper error library instead, the dependency is removed altogether moving to jest 29.