WIP Try supporting ESM with Jest 29
Jest 29 has multiple improvements, and in particular it adds support for the "browser" field in the package.json if using jsdom. However, jose ships as ESM-only for the browser module, so we must support ESM there. In any case, this means that we were not testing with the same code that gets deployed to our customers, which is a problem we must resolve no matter what.

Currently, this doesn't work, importing jose fails when running the oidc package test (the first package in the browser dependency chain).