Do not normalize the user-provided redirect URL
The redirect URL provided to the `login` function was being normalized, which
could result in misalignments with the redirect URLs declared in a Client Identifier
document, since the latter wasn't being normalized. The normalization step has
now been removed, and the redirect provided by the user is sent to the OpenID
Provider unchanged.