Fix token expiration on Client Credentials flow (#2658)
The Client Credential flow had a bug where the expiration time set by the OpenID Provider for the token was ignored, and an arbitrary default was applied instead. This resulted in the session being unable to make authenticated requests, but still acting as if it were logged in. The session now uses the expiration time set by the OpenID Provider.

Co-authored-by: Nicolas Seydoux <nseydoux@inrupt.com>