Make client identifier configurable in test app
In order to be able to test that logging in with a Client Identifier works well, the test driver should be able to pass the client ID created by the fixtures to the main app. A clean way to do so is to add a text field to the test app so that the test runner can type the client id in.