Proxy from session.events to session
In order to prevent a breaking change, the Session class should keep extending the EventEmitter API for the time being. To keep things simple, session.events will proxy to session, so that even during the transition there is a single EventEmitter instance. When we decide to make the breaking change, we can just remove the proxying, and have session.events be the EventEmitter instance instead of the session.