Clear all OIDC params from stored redirect URL
The redirect URL is stored on redirect to be reused in case of user refresh. However, it must be cleared from all OpenID params, and must not be altered compared to the originally registered redirect URL sent as part of the client's metadata. In particular, the client domain may not include a trailing slash, and this should not be normalized.