Run CI on multiple OS (#3115)
`Nx` (used by Lerna) has OS-specific dependencies, and them not being installed on windows breaks the toolchain. These dependencies have been added as optional so that they are in the package-lock, which is consumed by developers building the packages locally.

In order to catch such issues early, the CI now runs on all three major OS instead of just ubuntu. We moved away from running CI on non-ubuntu because of flakyness of the non-ubuntu runners, but in this case we would only be running unit tests, which are less flaky that end-to-end tests. The tests can be prevented from failing the whole workflow on fail in the future if they appear to be still flaky.