Fix bearer tokens support in browser
The current implementation relied on `oidc-client` in the browser to
perform part of the authorization grant. However, a bug caused the
RP authentication method to the OP token endpoint to default to
"client_secret_post", instead of "client_secret_basic", which is the
default used during dynamic client registration. This mismatch was
preventing the auth code grant to complete.

This grant is now handled by internal code, the same way the DPoP grant
is handled.