Allow to disable session background refresh
In a server-side scenario, one may not want sessions to proactively
refresh to keep their access token active. Instead, the session si
logged in using the refresh token when loaded from storage, but it will
only be active during the lifetime of its Access Token.

When building a Session, a new option "keepAlive" is supported. It
defaults to 'true' for backwards compatibility. If set to 'false', the
refresh token of the session will not be used to proactively refresh the
access token. `keepAlive` is persisted in storage with the session state
so that it is used after the authorization request during the token
request in the authorization code flow.