Handle dynamic client expiration
Dynamically registered clients eventually expire, after a duration
provided by the OpenID Provider at registration time. This change keeps
track of the expiration date in the browser code, and performs a now
dynamic client registration if the client in use expired.

If a legacy client is present in storage, it will be rotated as if it
were expired, to avoid the issue where an expired client is used and
results in an error on the OP.