Fix race condition in browser-based test
In some cases, a stale value was read from the application UI wile it
was being re-rendered. This change makes it so that the test-id being
rendered is removed from the DOM while the data is being fetched so that
the test script waits for the latest value to be displayed.