Keep refreshed tokens type consistent (#4103)
The `refreshTokens` function no longer ignores the refreshed token type, and keeps the token type consistent: a DPoP-bound token as an input will result in a refreshed DPoP-bound token (using the same DPoP key), and a non-DPoP bound token (i.e. a `Bearer` token) as an input will result in a refreshed token that is not DPoP-bound.